Move unlockers interfaces from game bundle into component
